#7346 Summary Plot: Improve default curve for observed data
