+++ ApplicationLibCode/Commands/SummaryPlotCommands/CMakeLists_files.cmake
@@ -1,6 +1,7 @@
 
 set (SOURCE_GROUP_HEADER_FILES
 ${CMAKE_CURRENT_LIST_DIR}/RicNewSummaryPlotFeature.h
+${CMAKE_CURRENT_LIST_DIR}/RicNewDefaultSummaryPlotFeature.h
 ${CMAKE_CURRENT_LIST_DIR}/RicNewSummaryCrossPlotFeature.h
 ${CMAKE_CURRENT_LIST_DIR}/RicNewSummaryCurveFeature.h
 ${CMAKE_CURRENT_LIST_DIR}/RicSummaryPlotFeatureImpl.h
@@ -47,6 +48,7 @@ ${CMAKE_CURRENT_LIST_DIR}/RicNewDerivedSummaryFeature.h
 
 set (SOURCE_GROUP_SOURCE_FILES
 ${CMAKE_CURRENT_LIST_DIR}/RicNewSummaryPlotFeature.cpp
+${CMAKE_CURRENT_LIST_DIR}/RicNewDefaultSummaryPlotFeature.cpp
 ${CMAKE_CURRENT_LIST_DIR}/RicNewSummaryCrossPlotFeature.cpp
 ${CMAKE_CURRENT_LIST_DIR}/RicNewSummaryCurveFeature.cpp
 ${CMAKE_CURRENT_LIST_DIR}/RicSummaryPlotFeatureImpl.cpp
