+++ ApplicationLibCode/Commands/SummaryPlotCommands/RicNewSummaryPlotFeature.cpp
@@ -147,154 +147,3 @@ void RicNewSummaryPlotFeature::setupActionLook( QAction* actionToSetup )
     actionToSetup->setText( "Open Summary Plot Editor" );
     actionToSetup->setIcon( QIcon( ":/SummaryPlotLight16x16.png" ) );
 }
-
-//==================================================================================================
-///
-//==================================================================================================
-
-#include "RicNewSummaryCurveFeature.h"
-#include "RicSummaryPlotFeatureImpl.h"
-#include "RimMainPlotCollection.h"
-#include "RiuPlotMainWindowTools.h"
-
-CAF_CMD_SOURCE_INIT( RicNewDefaultSummaryPlotFeature, "RicNewDefaultSummaryPlotFeature" );
-
-void extractPlotObjectsFromSelection( std::vector<RimSummaryCase*>*           selectedIndividualSummaryCases,
-                                      std::vector<RimSummaryCaseCollection*>* selectedEnsembles )
-{
-    CAF_ASSERT( selectedIndividualSummaryCases && selectedEnsembles );
-
-    // First try selected ensembles
-    caf::SelectionManager::instance()->objectsByTypeStrict( selectedEnsembles );
-    if ( !selectedEnsembles->empty() )
-    {
-        return;
-    }
-    // Second try selected summary cases
-    caf::SelectionManager::instance()->objectsByTypeStrict( selectedIndividualSummaryCases );
-    if ( !selectedIndividualSummaryCases->empty() )
-    {
-        return;
-    }
-
-    RimSummaryPlotCollection* sumPlotColl =
-        caf::SelectionManager::instance()->selectedItemAncestorOfType<RimSummaryPlotCollection>();
-
-    if ( sumPlotColl )
-    {
-        RimSummaryCase*           firstIndividualSummaryCase = nullptr;
-        RimSummaryCaseCollection* firstEnsemble              = nullptr;
-
-        auto sumCaseVector = RimProject::current()->allSummaryCases();
-        for ( RimSummaryCase* summaryCase : sumCaseVector )
-        {
-            RimSummaryCaseCollection* parentEnsemble = nullptr;
-            summaryCase->firstAncestorOrThisOfType( parentEnsemble );
-            if ( !parentEnsemble && !firstIndividualSummaryCase )
-            {
-                firstIndividualSummaryCase = summaryCase;
-                break;
-            }
-            else if ( parentEnsemble && !firstEnsemble )
-            {
-                firstEnsemble = parentEnsemble;
-            }
-        }
-        if ( firstIndividualSummaryCase )
-        {
-            selectedIndividualSummaryCases->push_back( firstIndividualSummaryCase );
-        }
-        else if ( firstEnsemble )
-        {
-            selectedEnsembles->push_back( firstEnsemble );
-        }
-    }
-}
-
-//--------------------------------------------------------------------------------------------------
-///
-//--------------------------------------------------------------------------------------------------
-RimSummaryPlot* RicNewDefaultSummaryPlotFeature::createFromSummaryCases( RimSummaryPlotCollection* plotCollection,
-                                                                         const std::vector<RimSummaryCase*>& summaryCases )
-{
-    RimSummaryPlot* newPlot = plotCollection->createSummaryPlotWithAutoTitle();
-
-    for ( RimSummaryCase* sumCase : summaryCases )
-    {
-        RicSummaryPlotFeatureImpl::addDefaultCurvesToPlot( newPlot, sumCase );
-    }
-
-    newPlot->applyDefaultCurveAppearances();
-    newPlot->loadDataAndUpdate();
-
-    plotCollection->updateConnectedEditors();
-
-    RiuPlotMainWindowTools::setExpanded( newPlot );
-    RiuPlotMainWindowTools::selectAsCurrentItem( newPlot );
-    return newPlot;
-}
-
-//--------------------------------------------------------------------------------------------------
-///
-//--------------------------------------------------------------------------------------------------
-bool RicNewDefaultSummaryPlotFeature::isCommandEnabled()
-{
-    std::vector<RimSummaryCase*>           selectedIndividualSummaryCases;
-    std::vector<RimSummaryCaseCollection*> selectedEnsembles;
-
-    extractPlotObjectsFromSelection( &selectedIndividualSummaryCases, &selectedEnsembles );
-
-    RimCustomObjectiveFunctionCollection* customObjFuncCollection = nullptr;
-    RimEnsembleCurveFilter*               curveFilter             = nullptr;
-    caf::PdmObject* selObj = dynamic_cast<caf::PdmObject*>( caf::SelectionManager::instance()->selectedItem() );
-    if ( selObj )
-    {
-        selObj->firstAncestorOrThisOfType( customObjFuncCollection );
-        selObj->firstAncestorOrThisOfType( curveFilter );
-    }
-    if ( customObjFuncCollection || curveFilter ) return false;
-
-    return !( selectedIndividualSummaryCases.empty() && selectedEnsembles.empty() );
-}
-
-//--------------------------------------------------------------------------------------------------
-///
-//--------------------------------------------------------------------------------------------------
-void RicNewDefaultSummaryPlotFeature::onActionTriggered( bool isChecked )
-{
-    std::vector<RimSummaryCase*>           selectedIndividualSummaryCases;
-    std::vector<RimSummaryCaseCollection*> selectedEnsembles;
-    extractPlotObjectsFromSelection( &selectedIndividualSummaryCases, &selectedEnsembles );
-
-    if ( !selectedIndividualSummaryCases.empty() )
-    {
-        RimSummaryPlotCollection* sumPlotColl = RimProject::current()->mainPlotCollection()->summaryPlotCollection();
-        createFromSummaryCases( sumPlotColl, selectedIndividualSummaryCases );
-    }
-    else
-    {
-        CAF_ASSERT( !selectedEnsembles.empty() );
-        RicNewSummaryEnsembleCurveSetFeature::createPlotForCurveSetsAndUpdate( selectedEnsembles );
-    }
-}
-
-//--------------------------------------------------------------------------------------------------
-///
-//--------------------------------------------------------------------------------------------------
-void RicNewDefaultSummaryPlotFeature::setupActionLook( QAction* actionToSetup )
-{
-    std::vector<RimSummaryCase*>           selectedIndividualSummaryCases;
-    std::vector<RimSummaryCaseCollection*> selectedEnsembles;
-
-    extractPlotObjectsFromSelection( &selectedIndividualSummaryCases, &selectedEnsembles );
-
-    if ( !selectedIndividualSummaryCases.empty() )
-    {
-        actionToSetup->setText( "New Summary Plot" );
-    }
-    else
-    {
-        actionToSetup->setText( "New Ensemble Summary Plot" );
-    }
-    actionToSetup->setIcon( QIcon( ":/SummaryPlotLight16x16.png" ) );
-}
