+++ ApplicationLibCode/ProjectDataModelCommands/RimcSummaryPlotCollection.cpp
@@ -15,8 +15,10 @@
 //  for more details.
 //
 /////////////////////////////////////////////////////////////////////////////////
+
 #include "RimcSummaryPlotCollection.h"
 
+#include "SummaryPlotCommands/RicNewDefaultSummaryPlotFeature.h"
 #include "SummaryPlotCommands/RicNewSummaryEnsembleCurveSetFeature.h"
 #include "SummaryPlotCommands/RicNewSummaryPlotFeature.h"
 #include "SummaryPlotCommands/RicSummaryPlotFeatureImpl.h"
